requirements.txt
scrapy==2.11.2
selenium==4.17.2
beautifulsoup4==4.12.3
requests==2.31.0
apscheduler==3.10.4
instaloader==4.9.6
python-telegram-bot==20.5
python-dotenv==1.0.0
python-dateutil==2.8.2
pyyaml==6.0.1
email-validator==2.1.0
tqdm==4.66.2
pytz==2024.1
